normalize

+++ docs/tools/translate/normalize-en-markdown.sh
@@ -5,5 +5,8 @@ BASE_DIR=$(dirname $(readlink -f $0))
 TEMP_FILE=$(mktemp)
 trap 'rm -f -- "${TEMP_FILE}"' INT TERM HUP EXIT
 INPUT="$1"
-cat "${INPUT}" > "${TEMP_FILE}"
-"${BASE_DIR}/translate.sh" "en" "${TEMP_FILE}" "${INPUT}"
+if [[ ! -L "${INPUT}" ]]
+then
+    cat "${INPUT}" > "${TEMP_FILE}"
+    "${BASE_DIR}/translate.sh" "en" "${TEMP_FILE}" "${INPUT}"
+fi
